Extract cloud authentication
Extract auth logic to a concern,
so it can be easily used in multiple places.
