+++ lib/foreman_rh_cloud/engine.rb
@@ -13,6 +13,7 @@ module ForemanRhCloud
     config.autoload_paths += Dir["#{config.root}/app/helpers/concerns"]
     config.autoload_paths += Dir["#{config.root}/app/models/concerns"]
     config.autoload_paths += Dir["#{config.root}/app/overrides"]
+    config.autoload_paths += Dir["#{config.root}/app/services"]
     config.autoload_paths += Dir["#{config.root}/lib"]
 
     config.eager_load_paths += Dir["#{config.root}/lib"]
